requirements.dev.txt
annotated-types==0.5.0
APScheduler==3.6.3
attrs==22.1.0
beautifulsoup4==4.11.1
black==23.7.0
bs4==0.0.1
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
clarifai-grpc==8.11.0
click==8.1.6
dnspython==2.2.1
emoji==2.2.0
ephem==4.1.3
exceptiongroup==1.0.4
fake-useragent==1.1.1
flake8==6.0.0
googleapis-common-protos==1.56.4
grpcio==1.51.1
idna==3.4
iniconfig==1.1.1
isort==5.12.0
mccabe==0.7.0
mongomock==4.1.2
mypy-extensions==1.0.0
numpy==1.24.1
packaging==22.0
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.9.1
pluggy==1.0.0
protobuf==3.19.3
pycodestyle==2.10.0
pydantic==1.10.7
pydantic_core==2.1.2
pyflakes==3.0.1
pymongo==4.3.3
PySocks==1.7.1
pytest==7.2.0
pytest-mongodb==2.2.0
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
sentinels==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tornado==6.1
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
